added view-listing component and browse component.
Browse component is nowhere near done, it's just placeholder and gets the job done ugly.
view-listing component is made to provide /listings/id destination. Now we have a page for each individual listing according to their unique ID. The issue now is that when creating a new listing, it doesn't automatically take you to that page. I think it's because it's asynch calls, it's not waiting until the post gets accepted into the db before redirecting me to the /listings/id url